Make creating a new section not dump the user on a blank page
